+++ resources/assets/es6/maps.js
@@ -3,6 +3,6 @@ import addMap from './mapbox-utils';
 
 let mapDivs = document.querySelectorAll('.map');
 
-for (var div of mapDivs) {
+for (let div of mapDivs) {
     addMap(div);
 }