+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Version {next}
   - Improve CSP headers
+  - Fix the map style menu in mapbox maps
 
 ## Version 0.16.3 (2018-04-12)
   - Improve JSON feed conformance